requirements.txt
certifi==2018.8.24
chardet==3.0.4
idna==2.7
mysql-connector==2.1.6
pkg-resources==0.0.0
requests==2.19.1
urllib3==1.23
